feat: adds safe transaction helper function
Introduces a `safe_transaction` helper function to wrap database
actions within a transaction. It verifies if the number of records
affected matches the expected row count of a reference object,
committing the transaction if it aligns and rolling back otherwise.
This change also converts the `populate_database.R` script to
`populate_database.qmd` to allow for easier documentation.